requirements.txt
aiodns==3.0.0
aiofile==3.5.0
aiohttp==3.7.4.post0
amqp==2.6.1
apns2==0.7.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
boto3==1.17.3
botocore==1.20.3
caio==0.8.0
celery==4.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.6
defusedxml==0.7.0rc2
Django==3.1.6
django-braces==1.14.0
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==2.4.0
django-oauth-toolkit==1.4.0
django-push-notifications==2.0.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.11.1
django-timezone-field==4.2.1
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
docutils==0.15.2
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
ibm-cloud-sdk-core==3.2.0
ibm-cos-sdk==2.7.0
ibm-cos-sdk-core==2.7.0
ibm-cos-sdk-s3transfer==2.7.0
idna==2.10
importlib-metadata==3.4.0
jmespath==0.10.0
kombu==4.6.11
Markdown==3.3.3
multidict==5.1.0
mysqlclient==2.0.3
numpy==1.21.1
oauthlib==3.1.0
Pillow==8.1.0
py-vapid==1.8.2
pycares==4.0.0
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pywebpush==1.13.0
redis==3.5.3
rejson==0.5.4
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sorl-thumbnail==12.7.0
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.4.1
stream-chat==3.7.0
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==1.3.0
yarl==1.6.3
zipp==3.4.0
